Added code to clean .lock file of MongoDB
Many time it happens that Mongo is shut down in a badway by Windows Shutdown or for any other reason. This cause the next start of MongoDB to fail and each time i have to go in that folder to delete that lock file.

This additive code check if it is present and delete it before starting Mongo.